requirements.txt
numpy==1.25.2
Pillow==10.0.0
progressbar2==4.2.0
python-utils==3.7.0
typing_extensions==4.7.1
